+++ src/main/java/repos/IWorkspaceRepository.java
@@ -3,5 +3,5 @@ package repos;
 import Models.Workspace;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IWorkspaceRepository extends JpaRepository<Workspace, Integer> {
+public interface IWorkspaceRepository extends JpaRepository<Workspace, Long> {
 }
